Added files via upload
In this approach, first I created a KML file from the field map (.CSV) using Python, generated polygon shapefiles with known size for each plot from KML file, and assigned plot ID to each plot using the field map. To define the geographic extent of the field, the python script asks for the coordinate of two points in the field: the start point of the first plot on the top right and the end point of the last plot on the bottom left. Then it starts from the top right and builds the first polygon using the defined plot size, and skips the gap between plots and generates the next one until it gets to the last plot on the bottom left. In this approach the plot IDs are assigned automatically and simultaneously from the field map excel sheet.